% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_rank_dev3.R
\name{createGeneRank}
\alias{createGeneRank}
\title{Create Gene Rank}
\usage{
createGeneRank(resultDF, value.col, names.col)
}
\arguments{
\item{resultDF}{result dataframe (DESeq2, edgeR)}

\item{value.col}{column as character to use as value in gene rank vector (eg. logFC, stat)}

\item{names.col}{column as character to use as names of gene rank vector (eg. entrez, ensembl)}
}
\value{
vector of gene rank
}
\description{
Create gene rank scores from differential analysis result data frames
}
\examples{
\dontrun{
createGeneRank(resultDF, 'stat', 'entGene')
}
}
