% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_rank_dev3.R
\name{drawGeneCntCutoffPeak2}
\alias{drawGeneCntCutoffPeak2}
\title{Draw peak plot}
\usage{
drawGeneCntCutoffPeak2(peakObj, mtitle = "", lw_colname = NULL)
}
\arguments{
\item{peakObj}{peak result object. (Output from getPeakResults2)}

\item{mtitle}{main title for plot}

\item{lw_colname}{loess smoothed column name. (from names(peakObj$smoothed_peak))}
}
\description{
Draw peak plot
}
