% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_rank_dev3.R
\name{getGenesByCounts}
\alias{getGenesByCounts}
\title{Bin Genes by rank}
\usage{
getGenesByCounts(
  genesrank,
  max_deg_count = 2000,
  bin_size = 10,
  reverse = FALSE
)
}
\arguments{
\item{genesrank}{named vector of genes rank (output from get_gene_rank)}

\item{max_deg_count}{maximum count for DEG}

\item{bin_size}{gene rank bin size}

\item{reverse}{reverse order (for down regulated genes)}
}
\value{
list of genes by rank cut-off
}
\description{
Create a list of genes for each binned count
}
