% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazy2_genelist.R
\name{getGenesByCutoffs}
\alias{getGenesByCutoffs}
\title{Get DEGs by cut-offs}
\usage{
getGenesByCutoffs(resultDF, fcos, qcos)
}
\arguments{
\item{resultDF}{result data frame. Column names need to include "entGene", "adj.P.Val", "logFC"}

\item{fcos}{fold change}

\item{qcos}{q-value}
}
\value{
List of entrez IDs
}
\description{
Extract the list of DEGs by cut-offs from the differential analysis results.
}
