% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{getPeakResults2}
\alias{getPeakResults2}
\title{Get Peak Result from calculatePathwayScores}
\usage{
getPeakResults2(geneCntList, scoresMat, verbose = FALSE)
}
\arguments{
\item{geneCntList}{list of genes binned by count (output from binGenesByCntCutoff)}

\item{scoresMat}{Pathway score matrix (output from calculatePathwayScores)}

\item{verbose}{print peak scores. Default FALSE}
}
\value{
PIS object list
}
\description{
Summarize result from calculatePathwayScores and get peak cut-off.
}
