% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{hypergeo_test_geneset}
\alias{hypergeo_test_geneset}
\title{Hypergeometric test for gmt style list}
\usage{
hypergeo_test_geneset(
  query,
  refGMT,
  gspace,
  min_geneset = 10,
  ef.psc = 1,
  verbose = FALSE
)
}
\arguments{
\item{query}{gene set to query (eg. Differentially Expressed Genes)}

\item{refGMT}{list of reference gene set (eg. Pathways)}

\item{gspace}{background gene space. Should contain all genes in query.}

\item{min_geneset}{minimum size of gene set used to filter refGMT. Default 10}

\item{ef.psc}{pseudocount when calculating enrichment factor (oddsRatio). Default 1}

\item{verbose}{print number of filtered entries in refGMT. Default FALSE}
}
\value{
Data frame of gene set analysis results
\describe{
\item{ID}{: pathway description or ID}
\item{pVal}{: hypergeometric test p values from phyper}
\item{qVal}{: FDR adjusted p-values}
\item{oddsRatio}{: odds ratio}
\item{tan}{: tanimoto coefficient (Jaccard index)}
\item{int}{: intersected item count}
\item{gsRatio}{: gene set ratio (selected genes in gene set / selected genes)}
\item{bgRatio}{: background ratio (total genes in gene set / total gene space)}
}

dataframe of results
}
\description{
Gene list overrepresentation analysis against list of gene set.
Runs both hypergeometric test or calculate enrichment factor for gene set. (used inside calculatePathwayScores)
}
