% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{hypergeoTestForGeneset.simple}
\alias{hypergeoTestForGeneset.simple}
\title{Hypergeometric Test for Gene set}
\usage{
hypergeoTestForGeneset.simple(
  query,
  refGMT,
  gspace,
  minGeneSet = 10,
  ef.psc = 1,
  verbose = FALSE
)
}
\arguments{
\item{query}{gene set to query (eg. Differentially Expressed Genes)}

\item{refGMT}{list of reference gene set (eg. Pathways)}

\item{gspace}{background gene space. Should contain all genes in query.}

\item{minGeneSet}{minimum size of gene set. This is used to filter refGMT. Default 10}

\item{ef.psc}{pseudocount when calculating enrichment factor (oddsRatio). Default 1}

\item{verbose}{print number of filtered entries in refGMT. Default FALSE}
}
\description{
Run hypergeometric test for gene set. (used inside calculatePathwayScores)
This is a simplified version of the function in the Lazy2 package.
}
