% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{intg_pathways}
\alias{intg_pathways}
\alias{intg_pathinfo}
\title{Integrated pathway set}
\format{
A list of gene sets (human entrez ID). Pathway descriptions in intg_pathinfo.

A data.table describing pathways in 'intg_pathways'
}
\usage{
data(intg_pathways)

data(intg_pathinfo)
}
\description{
Integrated pathway set from KaiPharm. (sources from Reactome, GO)
}
\section{Functions}{
\itemize{
\item \code{intg_pathinfo}: Pathway information metadata for intg_pathways

}}
\examples{
data(intg_pathways)
data(intg_pathinfo)
}
\keyword{datasets}
