% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maqc_deseq2}
\alias{maqc_deseq2}
\title{MAQC2 deseq2 analysis result}
\format{
A dataframe of 16851 rows and 10 columns:
\describe{
\item{entrez}{NCBI Entrez ID, Used in ent2sym}
\item{hgnc_id}{HGNC ID}
\item{hgnc_gene}{HGNC gene name}
\item{hgnc_symbol}{HGNC approved gene symbol, Used in ent2sym}
\item{ensembl}{Ensembl ID}
\item{gene_type}{gene type}
}
}
\source{
\href{https://www.gencodegenes.org/human/}{Gencode Metadata}
}
\usage{
data(maqc_deseq2)
}
\description{
Human gene ID dataframe containing Entrez, HGNC ID, Gene name, Gene symbol, Ensembl ID, gene type
Used for ent2sym and more gene mapper functions.
}
\examples{
data(maqc_deseq2)
}
\keyword{datasets}
