% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISobj_class.R
\name{new_PISobj}
\alias{new_PISobj}
\title{Generate PIS object}
\usage{
new_PISobj(obj = NULL)
}
\arguments{
\item{obj}{input object to make into PISobj. Leaving this empty will generate an empty object.}
}
\description{
Generate an empty PIS object or turn a list into PISobj
}
