% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{pathScores2}
\alias{pathScores2}
\title{Pathway score}
\usage{
pathScores2(efs, ef_cut = 2)
}
\arguments{
\item{efs}{vector of enrichment factors}

\item{ef_cut}{minimum enrichment factor cut-offs. Default 2}
}
\value{
vector of pathway enrichment score
}
\description{
Calculate pathway scores from enrichment factors (used inside calculatePathwayScores)
}
