% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{pathway_scores}
\alias{pathway_scores}
\title{Pathway score}
\usage{
pathway_scores(efs, ef_cut = 2)
}
\arguments{
\item{efs}{vector of enrichment factors}

\item{ef_cut}{minimum enrichment factor cut-offs. Default 2}
}
\value{
vector of pathway scores
}
\description{
Pathway score
}
