% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_rank_dev3.R
\name{peakSignifByRandPerm}
\alias{peakSignifByRandPerm}
\title{Random permutation}
\usage{
peakSignifByRandPerm(
  peakObj,
  gspace,
  ef_cut = 2,
  min.overlap = 5,
  iter = 10000,
  use.smoothed = FALSE,
  ncore = 1
)
}
\arguments{
\item{peakObj}{peak result object. (Output from getPeakResults2)}

\item{gspace}{total gene space}

\item{ef_cut}{EF cut-off}

\item{min.overlap}{minimum overlap}

\item{iter}{iteration}

\item{use.smoothed}{use smoothed?}

\item{ncore}{number of cores}
}
\value{
peak result object with smoothed scores attribute
}
\description{
(add description)
}
