% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS-package.R
\docType{package}
\name{PIS-package}
\alias{PIS}
\alias{PIS-package}
\title{PIS: Pathway Impact Score}
\description{
Determine DEG thresholds by calculating the pathway level effects.
}
\author{
\strong{Maintainer}: Yeogha Yoon \email{ygyoon17@ewhain.net} (\href{https://orcid.org/0000-0003-0955-4344}{ORCID})

}
\keyword{internal}
