% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISobj_class.R
\name{print.PISobj}
\alias{print.PISobj}
\title{Printing PIS object}
\usage{
\method{print}{PISobj}(x, ...)
}
\arguments{
\item{x}{object of 'PISobj'}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print PIS object.
}
\examples{
# S3 method for class 'PISobj'
x <- new_PISobj()
print(x)
}
