% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazy2_genelist.R
\name{readGMT}
\alias{readGMT}
\alias{writeGMT}
\title{readGMT}
\usage{
readGMT(gmtfile, as.df = FALSE)

writeGMT(gmtfile, genelist, geneset_desc = "")
}
\arguments{
\item{gmtfile}{GMT file path}

\item{as.df}{Return as data frame?}

\item{genelist}{List of gene set. Should be un-nested level one named list.}

\item{geneset_desc}{Description meta information for gene set. Either length one or a named vector of same length as glist.}
}
\value{
Gene set dataframe of 2 column or a list of gene sets.
}
\description{
Read/write GMT file. A GMT file format is a tab delimited text file containing gene sets.
Each line in the GMT file should contain one gene set or pathway, delimited by tab. Genes should start from the 3rd field.
}
\section{Functions}{
\itemize{
\item \code{writeGMT()}: write gmt file for geneset list.

}}
\examples{
\dontrun{
readGMT("file/path.gmt")
}
}
