% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_rank_dev3.R
\name{scoreSmooth}
\alias{scoreSmooth}
\title{Smooth scores}
\usage{
scoreSmooth(peakObj, loess.span = 0.1)
}
\arguments{
\item{peakObj}{peak result object. (Output from getPeakResults2)}

\item{loess.span}{span option for loess}
}
\value{
peak result object with smoothed scores attribute
}
\description{
Fit loess smoothing on bin scores
}
