% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISobj_class.R
\name{validate_PISobj}
\alias{validate_PISobj}
\title{Validate PIS object}
\usage{
validate_PISobj(obj)
}
\arguments{
\item{obj}{PISobj}
}
\description{
Validate PIS object or check if a list can become PIS object.
}
