% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{calculatePathwayScores}
\alias{calculatePathwayScores}
\title{PIS calculation}
\usage{
calculatePathwayScores(
  glist,
  gspace,
  ref_geneset,
  ef_cut = 2,
  ef.psc = 1,
  ncore = 4,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{glist}{list of gene set (result from binGenesByCntCutoff)}

\item{gspace}{background gene space. (for hypergeoTestForGeneset.simple)}

\item{ref_geneset}{list of reference gene set (eg. Pathways)}

\item{ef_cut}{EF cut-off for scoring.}

\item{ef.psc}{pseudocount when calculating enrichment factor (oddsRatio). Default 0}

\item{ncore}{no of threads to use in mclapply. (ncore=1 uses lapply)}

\item{verbose}{print run time. Default FALSE}

\item{...}{futher arguments to be passed to the internal hypergeoTestForGeneset.simple function}
}
\value{
Matrix of pathway scores
}
\description{
Calculate pathway score for gene list (uses pathScores2 function)
Outputs matrix of pathway scores where each row is gene set and column is bin.
}
