% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lazy2_genelist.R
\name{getGenesByCutoffs}
\alias{getGenesByCutoffs}
\title{Get DEGs by cut-offs}
\usage{
getGenesByCutoffs(
  resultDF,
  fcos,
  qcos,
  colname.qv = "adj.P.Val",
  colname.lfc = "logFC",
  colname.gene = "entGene"
)
}
\arguments{
\item{resultDF}{result data frame. Column names should include "entGene", "adj.P.Val", "logFC"}

\item{fcos}{sequence of fold changes to use as cut-offs}

\item{qcos}{sequence of q-values (FDR adjusted p-value) to use as cut-offs}

\item{colname.qv}{name for the column in resultDF containing q-values}

\item{colname.lfc}{name for the column in resultDF containing log2 Fold changes}

\item{colname.gene}{name for the column in resultDF containing the gene ID (eg. entrez, ensemble, symbol)}
}
\value{
List of gene IDs selected by cutoffs
}
\description{
Extract the list of DEGs by cut-offs from the differential analysis results.
}
\examples{
\dontrun{
data(maqc_deseq2)
fcos <- seq(1.5,4.0,0.1)
qcos <- seq(0.01,0.2, 0.01)
geneList.conv <- getGenesByCutoffs(maqc_deseq2, fcos, qcos)
# head(lapply(geneList.conv$up, head))
}

}
