% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{getGenesByCutoffs}
\alias{getGenesByCutoffs}
\title{Gene List Extraction}
\usage{
getGenesByCutoffs(resultDF, fcos, qcos)
}
\arguments{
\item{resultDF}{result data frame. column names need to include 'entGene', 'adj.P.Val', 'logFC'}

\item{fcos}{fold change}

\item{qcos}{q-value}
}
\value{
List of entrez IDs
}
\description{
Extract gene list by cut-offs from DE result
}
