% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIS_functions_dev3.R
\name{getPeakResults2}
\alias{getPeakResults2}
\title{Get Peak Result object from calculatePathwayScores}
\usage{
getPeakResults2(genelist, scores_mat, verbose = FALSE)
}
\arguments{
\item{genelist}{list of genes binned by cut-off.}

\item{scores_mat}{Pathway score matrix (output from calculatePathwayScores)}

\item{verbose}{print peak scores. Default FALSE}
}
\value{
PISobj with peak results.
\describe{
\item{peak_cnt}{: No. of DEGs selected}
\item{peak_score}{: max PIS score}
\item{peak_pathwayCnt}{: No. of pathways selected at peak}
\item{peak_gset}{: DEGs selected at peak}
\item{scored_pathways}{: Pathways and scores selected by PIS}
\item{bin_scores}{: scores by each cut-off)}
}
}
\description{
Summarize result from calculatePathwayScores and get peak cut-off.
}
